Paspportrait/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Django==5.2.1
dlib==19.24.8
face_recognition_models==0.3.0
face-recognition==1.3.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==24.3.1
platformdirs==4.3.8
pooch==1.8.2
PyMatting==1.1.13
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.3
sqlparse==0.5.3
tifffile==2025.5.10
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
